+++ backend/eodi-api/src/main/java/com/bb/eodi/deal/presentation/dto/request/RegionRecommendRequest.java
@@ -14,7 +14,7 @@ public record RegionRecommendRequest(
         @NotNull(message = "보유 현금은 필수값입니다.")
         @Positive(message = "보유 현금은 0보다 커야합니다.")
         @Parameter(description = "보유 현금 (단위: 만 원)", example = "24000")
-        Integer cash,
+        Long cash,
 
         @Parameter(description = "대출 여부", example = "true")
         Boolean hasLoan,
@@ -23,10 +23,10 @@ public record RegionRecommendRequest(
         Boolean isFirstTimeBuyer,
 
         @Parameter(description = "대출 월 상환액 (단위: 만 원)", example = "210")
-        Integer monthlyPayment,
+        Long monthlyPayment,
 
         @Parameter(description = "연 소득 (단위: 만 원)", example = "5000")
-        Integer annualIncome,
+        Long annualIncome,
 
         @Parameter(description = "대상 주택 유형 목록", example = "AP")
         List<String> housingTypes
