+++ backend/eodi-api/src/main/java/com/bb/eodi/finance/application/service/MortgageLoanCalculateService.java
@@ -37,7 +37,7 @@ public class MortgageLoanCalculateService {
 
     }
 
-    public int calculateMaximumMortgageLoanAmount(int cash) {
+    public long calculateMaximumMortgageLoanAmount(long cash) {
         return (cash * mortgagePolicy.getDefaultLtv()) / (100 - mortgagePolicy.getDefaultLtv());
     }
 }
